metric/python/requirements.txt
flask==3.0.0
requests==2.31.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-propagator-b3==1.21.0
opentelemetry-exporter-prometheus==1.12.0rc1
